Move source files into src, change import paths, fixing Makefile
