+++ Makefile
@@ -1,27 +1,52 @@
-ifeq ($(GOPATH),)
-export GOPATH=/tmp/go
+all: precheck deps telemetry
+GO=go
+SRC_FILES=$(wildcard *.go)
+TOP_DIR := $(abspath ../..)
+GOFLAGS:=
+BUILD_DIR=build
+GO_DEP_PATH=$(abspath .)/$(BUILD_DIR)
+GO_MGMT_PATH=$(TOP_DIR)/sonic-mgmt-framework
+GO_SONIC_TELEMETRY_PATH=$(TOP_DIR)
+CVL_GOPATH=$(GO_MGMT_PATH)
+GOPATH = $(CVL_GOPATH):$(GO_DEP_PATH):$(GO_MGMT_PATH):$(GO_SONIC_TELEMETRY_PATH)
+
+ifdef DEBUG
+	GOFLAGS += -gcflags="all=-N -l"
 endif
 
-INSTALL := /usr/bin/install
+precheck:
+	$(shell mkdir -p $(BUILD_DIR))
 
-all: sonic-telemetry
+deps: $(BUILD_DIR)/.deps
 
-sonic-telemetry:
-	/usr/local/go/bin/go get -v github.com/Azure/sonic-telemetry/telemetry
-	/usr/local/go/bin/go get -v github.com/Azure/sonic-telemetry/dialout/dialout_client_cli
+$(BUILD_DIR)/.deps:
+	touch $(BUILD_DIR)/.deps
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u  github.com/Workiva/go-datastructures/queue
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/openconfig/goyang
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/openconfig/ygot/ygot
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/golang/glog
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/go-redis/redis
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u  github.com/c9s/goprocinfo/linux
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u  github.com/golang/protobuf/proto
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u  github.com/openconfig/gnmi/proto/gnmi
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u  golang.org/x/net/context
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u  google.golang.org/grpc
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u google.golang.org/grpc/credentials
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u gopkg.in/go-playground/validator.v9
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/gorilla/mux
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/openconfig/goyang
+	GOPATH=$(GO_DEP_PATH) $(GO) get -u github.com/openconfig/ygot/ygot
 
-check:
-	/usr/local/go/bin/go get -v -t github.com/Azure/sonic-telemetry/gnmi_server/...
-	-/usr/local/go/bin/go test -v ${GOPATH}/src/github.com/Azure/sonic-telemetry/gnmi_server
+telemetry:$(BUILD_DIR)/telemetry
 
-install:
-	$(INSTALL) -D ${GOPATH}/bin/telemetry $(DESTDIR)/usr/sbin/telemetry
-	$(INSTALL) -D ${GOPATH}/bin/dialout_client_cli $(DESTDIR)/usr/sbin/dialout_client_cli
-
-deinstall:
-	rm $(DESTDIR)/usr/sbin/telemetry
-	rm $(DESTDIR)/usr/sbin/dialout_client_cli
+$(BUILD_DIR)/telemetry:
+	@echo "Building $@"
+	GOPATH=$(GOPATH) $(GO) build $(GOFLAGS) -o $@ $(SRC_FILES)
 
 clean:
-	rm -fr ${GOPATH}
+	rm -rf $(BUILD_DIR)/telemetry
+	make -C  $(GO_MGMT_PATH) clean
 
+cleanall:
+	rm -rf $(BUILD_DIR)
+	make -C  $(GO_MGMT_PATH) cleanall
