+++ src/dialout/dialout_client_cli/dialout_client_cli.go
@@ -4,7 +4,7 @@ package main
 import (
 	"crypto/tls"
 	"flag"
-	dc "github.com/Azure/sonic-telemetry/dialout/dialout_client"
+	dc "dialout/dialout_client"
 	log "github.com/golang/glog"
 	gpb "github.com/openconfig/gnmi/proto/gnmi"
 	"golang.org/x/net/context"
