+++ src/dialout/dialout_server/dialout_server.go
@@ -3,7 +3,7 @@ package dialout_server
 import (
 	"errors"
 	"fmt"
-	spb "github.com/Azure/sonic-telemetry/proto"
+	spb "proto"
 	log "github.com/golang/glog"
 	"github.com/google/gnxi/utils"
 	gpb "github.com/openconfig/gnmi/proto/gnmi"
