+++ src/gnmi_server/client_subscribe.go
@@ -7,12 +7,12 @@ import (
 	"sync"
 
 	log "github.com/golang/glog"
-	"github.com/workiva/go-datastructures/queue"
+	"github.com/Workiva/go-datastructures/queue"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 
-	//spb "github.com/Azure/sonic-telemetry/proto"
-	sdc "github.com/Azure/sonic-telemetry/sonic_data_client"
+	//spb "github.com/project-arlo/sonic-telemetry/proto"
+	sdc "sonic_data_client"
 	gnmipb "github.com/openconfig/gnmi/proto/gnmi"
 )
