+++ src/sonic_data_client/non_db_client.go
@@ -3,11 +3,11 @@ package client
 import (
 	"encoding/json"
 	"fmt"
-	spb "github.com/Azure/sonic-telemetry/proto"
+	spb "proto"
 	linuxproc "github.com/c9s/goprocinfo/linux"
 	log "github.com/golang/glog"
 	gnmipb "github.com/openconfig/gnmi/proto/gnmi"
-	"github.com/workiva/go-datastructures/queue"
+	"github.com/Workiva/go-datastructures/queue"
 	"sync"
 	"time"
 )
