+++ src/testdata/COUNTERS:Ethernet1.txt
