+++ src/testdata/COUNTERS:Ethernet68.txt
