+++ src/testdata/COUNTERS:Ethernet68:Queues_alias.txt
