+++ src/testdata/COUNTERS:Ethernet_wildcard.txt
