+++ src/testdata/COUNTERS:Ethernet_wildcard_PFC_7_RX_alias.txt
