+++ src/testdata/COUNTERS:Ethernet_wildcard_Queues_alias.txt
