+++ src/testdata/COUNTERS:Ethernet_wildcard_alias.txt
