+++ src/testdata/COUNTERS:oid:0x1500000000091f.txt
