+++ src/testdata/COUNTERS:oid:0x1500000000092a.txt
