+++ src/testdata/COUNTERS_PORT_ALIAS_MAP.txt
