+++ src/testdata/COUNTERS_PORT_NAME_MAP.txt
